Linters report card (#267)
* Apply `gofmt` to all files

* Apply `gci` to all files

* Apply `gofumpt` to all files

* Enable `goconst` linter

* Enable `unconvert` linter

* Add GO report card

+++ context/ngap_handler.go
@@ -21,7 +21,6 @@ func HandlePDUSessionResourceSetupResponseTransfer(b []byte, ctx *SMContext) (er
 	resourceSetupResponseTransfer := ngapType.PDUSessionResourceSetupResponseTransfer{}
 
 	err = aper.UnmarshalWithParams(b, &resourceSetupResponseTransfer, "valueExt")
-
 	if err != nil {
 		return err
 	}
@@ -61,7 +60,6 @@ func HandlePDUSessionResourceSetupUnsuccessfulTransfer(b []byte, ctx *SMContext)
 	resourceSetupUnsuccessfulTransfer := ngapType.PDUSessionResourceSetupUnsuccessfulTransfer{}
 
 	err = aper.UnmarshalWithParams(b, &resourceSetupUnsuccessfulTransfer, "valueExt")
-
 	if err != nil {
 		return err
 	}
@@ -117,7 +115,7 @@ func HandlePathSwitchRequestTransfer(b []byte, ctx *SMContext) error {
 				DLPDR.FAR.ForwardingParameters.OuterHeaderCreation = new(pfcpType.OuterHeaderCreation)
 				dlOuterHeaderCreation := DLPDR.FAR.ForwardingParameters.OuterHeaderCreation
 				dlOuterHeaderCreation.OuterHeaderCreationDescription = pfcpType.OuterHeaderCreationGtpUUdpIpv4
-				dlOuterHeaderCreation.Teid = uint32(teid)
+				dlOuterHeaderCreation.Teid = teid
 				dlOuterHeaderCreation.Ipv4Address = gtpTunnel.TransportLayerAddress.Value.Bytes
 				DLPDR.FAR.State = RULE_UPDATE
 				DLPDR.FAR.ForwardingParameters.PFCPSMReqFlags = new(pfcpType.PFCPSMReqFlags)
@@ -133,7 +131,6 @@ func HandlePathSwitchRequestSetupFailedTransfer(b []byte, ctx *SMContext) (err e
 	pathSwitchRequestSetupFailedTransfer := ngapType.PathSwitchRequestSetupFailedTransfer{}
 
 	err = aper.UnmarshalWithParams(b, &pathSwitchRequestSetupFailedTransfer, "valueExt")
-
 	if err != nil {
 		return err
 	}
@@ -146,7 +143,6 @@ func HandleHandoverRequiredTransfer(b []byte, ctx *SMContext) (err error) {
 	handoverRequiredTransfer := ngapType.HandoverRequiredTransfer{}
 
 	err = aper.UnmarshalWithParams(b, &handoverRequiredTransfer, "valueExt")
-
 	if err != nil {
 		return err
 	}
@@ -159,7 +155,6 @@ func HandleHandoverRequestAcknowledgeTransfer(b []byte, ctx *SMContext) (err err
 	handoverRequestAcknowledgeTransfer := ngapType.HandoverRequestAcknowledgeTransfer{}
 
 	err = aper.UnmarshalWithParams(b, &handoverRequestAcknowledgeTransfer, "valueExt")
-
 	if err != nil {
 		return err
 	}
